gcp: record accurate GPU metadata for accelerator shapes
The Compute API reports fractional G4 accelerators as a whole-number
count and omits per-GPU memory, overstating fractional shapes and
leaving newer Blackwell models incomplete.

Add the missing memory mappings and represent G4 1/8, 1/4, and 1/2
slices as exact fractional counts. This is latent while A4, A4X, and G4
pricing is suppressed; existing A2, A3, and G2 output is unchanged.

+++ scraper/gcp/scrape.go
@@ -569,7 +569,7 @@ func processGCPData(skus []SKU, pricing map[string]PriceInfo, machineSpecs map[s
 			PrettyName:         createPrettyName(instanceType),
 			NetworkPerformance: "Variable",
 			Generation:         "current",
-			GPU:                float64(specs.GPU),
+			GPU:                specs.GPU,
 			GPUModel:           gpuModel,
 			GPUMemory:          specs.GPUMemory,
 			Pricing:            make(map[Region]map[OS]any),
